+++ pykilosort/gui/__init__.py
@@ -1,0 +1,10 @@
+"""
+This module contains all files required by the pykilosort GUI.
+"""
+from .data_view_box import DataViewBox
+from .probe_view_box import ProbeViewBox
+from .settings_box import SettingsBox
+from .message_log_box import MessageLogBox
+from .run_box import RunBox
+from .header_box import HeaderBox
+from .palettes import DarkPalette
