+++ pykilosort/gui/settings_box.py
@@ -1,30 +1,7 @@
 import os
-import sys
-import pyqtgraph as pg
-# TODO: optimize imports before incorporating into codebase
+from PyQt5 import QtWidgets, QtGui
+
 from pykilosort.default_params import default_params
-from PyQt5 import QtGui, QtWidgets, QtCore
-
-
-class HeaderBox(QtWidgets.QWidget):
-
-    def __init__(self, parent):
-        QtWidgets.QWidget.__init__(self, parent=parent)
-
-        self.layout = QtWidgets.QHBoxLayout()
-
-        self.kilosort_text = QtWidgets.QLabel()
-        self.kilosort_text.setText("KiloSort")
-        self.kilosort_text.setFont(QtGui.QFont("Arial", 20, QtGui.QFont.Black))
-        self.help_button = QtWidgets.QPushButton("Help")
-        self.reset_gui_button = QtWidgets.QPushButton("Reset GUI")
-
-        self.layout.addWidget(self.kilosort_text)
-        self.layout.addStretch(0)
-        self.layout.addWidget(self.help_button)
-        self.layout.addWidget(self.reset_gui_button)
-
-        self.setLayout(self.layout)
 
 
 class SettingsBox(QtWidgets.QGroupBox):
@@ -285,167 +262,3 @@ class SettingsBox(QtWidgets.QGroupBox):
         except AssertionError:
             self.error_label.setText("Invalid input!\nCheck that 0 <= AUC split <= 1!")
             self.error_label.show()
-
-
-class ProbeViewBox(QtWidgets.QGroupBox):
-
-    def __init__(self, parent):
-        QtWidgets.QGroupBox.__init__(self, parent=parent)
-        self.setTitle("Probe View")
-
-        self.layout = QtWidgets.QVBoxLayout()
-
-        self.info_message = QtWidgets.QLabel("scroll to zoom, click to view channel,\nright click to disable channel")
-        self.info_message.setFont(QtGui.QFont("Arial", 8, QtGui.QFont.Black))
-        self.info_message.setAlignment(QtCore.Qt.AlignCenter)
-
-        self.probe_view = pg.PlotWidget()
-
-        self.layout.addWidget(self.info_message, 5)
-        self.layout.addWidget(self.probe_view, 95)
-
-        self.setLayout(self.layout)
-
-
-class DataViewBox(QtWidgets.QGroupBox):
-
-    def __init__(self, parent):
-        QtWidgets.QGroupBox.__init__(self, parent=parent)
-        self.setTitle("Data View")
-
-        self.layout = QtWidgets.QVBoxLayout()
-
-        self.controls_button_layout = QtWidgets.QHBoxLayout()
-        self.controls_button = QtWidgets.QPushButton("Controls")
-        self.controls_button.setSizePolicy(QtWidgets.QSizePolicy.Expanding, QtWidgets.QSizePolicy.Expanding)
-        self.controls_button_layout.addWidget(self.controls_button)
-
-        self.data_view_layout = QtWidgets.QHBoxLayout()
-        self.data_view_widget = pg.PlotWidget()
-        self.data_view_layout.addWidget(self.data_view_widget)
-
-        self.data_seek_layout = QtWidgets.QHBoxLayout()
-        self.data_seek_widget = pg.PlotWidget()
-        self.data_seek_layout.addWidget(self.data_seek_widget)
-
-        self.layout.addLayout(self.controls_button_layout, 3)
-        self.layout.addLayout(self.data_view_layout, 82)
-        self.layout.addLayout(self.data_seek_layout, 15)
-
-        self.setLayout(self.layout)
-
-
-class RunBox(QtWidgets.QGroupBox):
-
-    def __init__(self, parent):
-        QtWidgets.QGroupBox.__init__(self, parent=parent)
-        self.setTitle("Run")
-
-        self.layout = QtWidgets.QGridLayout()
-
-        self.run_all_button = QtWidgets.QPushButton("Run All")
-        self.run_all_button.setSizePolicy(QtWidgets.QSizePolicy.Expanding, QtWidgets.QSizePolicy.Expanding)
-        self.preprocess_button = QtWidgets.QPushButton("Preprocess")
-        self.spike_sort_button = QtWidgets.QPushButton("Spikesort")
-        self.export_button = QtWidgets.QPushButton("Export for Phy")
-
-        self.layout.addWidget(self.run_all_button, 0, 0, 3, 2)
-        self.layout.addWidget(self.preprocess_button, 0, 2, 1, 2)
-        self.layout.addWidget(self.spike_sort_button, 1, 2, 1, 2)
-        self.layout.addWidget(self.export_button, 2, 2, 1, 2)
-
-        self.setLayout(self.layout)
-
-
-class MessageLogBox(QtWidgets.QGroupBox):
-
-    def __init__(self, parent):
-        QtWidgets.QGroupBox.__init__(self, parent=parent)
-        self.setTitle("Message Log")
-
-        self.layout = QtWidgets.QHBoxLayout()
-        self.log_box = QtWidgets.QPlainTextEdit()
-        self.layout.addWidget(self.log_box)
-
-        self.setLayout(self.layout)
-
-
-class KiloSortGUI(QtWidgets.QMainWindow):
-
-    def __init__(self, *args, **kwargs):
-        QtWidgets.QMainWindow.__init__(self, *args, **kwargs)
-
-        self.setWindowTitle("KiloSort2 GUI")
-        self.content = QtWidgets.QWidget(self)
-        self.content_layout = QtWidgets.QVBoxLayout()
-
-        self.header_box = HeaderBox(self)
-        self.content_layout.addWidget(self.header_box, 3)
-
-        self.boxes = QtWidgets.QWidget()
-        self.boxes_layout = QtWidgets.QHBoxLayout(self.boxes)
-        self.second_boxes_layout = QtWidgets.QVBoxLayout()
-
-        self.settings_box = SettingsBox(self)
-        self.probe_view_box = ProbeViewBox(self)
-        self.data_view_box = DataViewBox(self)
-        self.run_box = RunBox(self)
-        self.message_log_box = MessageLogBox(self)
-
-        self.second_boxes_layout.addWidget(self.settings_box, 85)
-        self.second_boxes_layout.addWidget(self.run_box, 15)
-
-        self.boxes_layout.addLayout(self.second_boxes_layout, 20)
-        self.boxes_layout.addWidget(self.probe_view_box, 15)
-        self.boxes_layout.addWidget(self.data_view_box, 65)
-
-        self.boxes.setLayout(self.boxes_layout)
-        self.content_layout.addWidget(self.boxes, 90)
-        self.content_layout.addWidget(self.message_log_box, 7)
-
-        self.content_layout.setContentsMargins(10, 10, 10, 10)
-        self.content.setLayout(self.content_layout)
-        self.setCentralWidget(self.content)
-
-
-class DarkPalette(QtGui.QPalette):
-    """Class that inherits from pyqtgraph.QtGui.QPalette and renders dark colours for the application."""
-    def __init__(self):
-
-        QtGui.QPalette.__init__(self)
-        self.setup()
-
-    def setup(self):
-
-        self.setColor(QtGui.QPalette.Window, QtGui.QColor(53, 50, 47))
-        self.setColor(QtGui.QPalette.WindowText, QtGui.QColor(255, 255, 255))
-        self.setColor(QtGui.QPalette.Base, QtGui.QColor(30, 27, 24))
-        self.setColor(QtGui.QPalette.AlternateBase, QtGui.QColor(53, 50, 47))
-        self.setColor(QtGui.QPalette.ToolTipBase, QtGui.QColor(255, 255, 255))
-        self.setColor(QtGui.QPalette.ToolTipText, QtGui.QColor(255, 255, 255))
-        self.setColor(QtGui.QPalette.Text, QtGui.QColor(255, 255, 255))
-        self.setColor(QtGui.QPalette.Button, QtGui.QColor(53, 50, 47))
-        self.setColor(QtGui.QPalette.ButtonText, QtGui.QColor(255, 255, 255))
-        self.setColor(QtGui.QPalette.BrightText, QtGui.QColor(255, 0, 0))
-        self.setColor(QtGui.QPalette.Link, QtGui.QColor(42, 130, 218))
-        self.setColor(QtGui.QPalette.Highlight, QtGui.QColor(42, 130, 218))
-        self.setColor(QtGui.QPalette.HighlightedText, QtGui.QColor(0, 0, 0))
-
-
-if __name__ == "__main__":
-    kilosort_application = QtWidgets.QApplication(sys.argv)
-    kilosort_application.setStyle("Fusion")
-    kilosort_application.setPalette(DarkPalette())
-    kilosort_application.setStyleSheet("QToolTip { color: #aeadac;"
-                                       "background-color: #35322f;"
-                                       "border: 1px solid #aeadac; }")
-
-    pg.setConfigOption('background', 'k')
-    pg.setConfigOption('foreground', 'w')
-    pg.setConfigOption('useOpenGL', True)
-
-    kilosort_gui = KiloSortGUI()
-    kilosort_gui.showMaximized()
-    kilosort_gui.show()
-
-    sys.exit(kilosort_application.exec_())
